프로젝트명 변경
프로젝트명 및 디렉토리명 변경
testproject -> wakeupAPI
