+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sun Feb 23 20:58:38 KST 2025
+gradle.version=8.12.1
