+++ .idea/material_theme_project_new.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="MaterialThemeProjectNewConfig">
+    <option name="metadata">
+      <MTProjectMetadataState>
+        <option name="userId" value="-7ffc926d:19532ac2a7c:-7ff0" />
+      </MTProjectMetadataState>
+    </option>
+  </component>
+</project>
