+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'testproject'
+rootProject.name = 'wakeupAPI'
