--- src/main/java/com/example/testproject/TestprojectApplication.java
@@ -1,13 +1,0 @@
-package com.example.testproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestprojectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TestprojectApplication.class, args);
-	}
-
-}
