+++ src/main/java/com/example/wakeupAPI/TestprojectApplication.java
@@ -1,0 +1,13 @@
+package com.example.wakeupAPI;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestprojectApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TestprojectApplication.class, args);
+	}
+
+}
