--- src/test/java/com/example/testproject/TestprojectApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.testproject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TestprojectApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
