+++ src/test/java/com/example/wakeupAPI/TestprojectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.wakeupAPI;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TestprojectApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
